+++ SpatialGDK/Source/SpatialGDK/Public/Utils/InterestFactory.h
